+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/domain/repository/SecretRepository.kt
@@ -15,7 +15,7 @@ interface SecretRepository {
 
     suspend fun createSecretCapsule (request: SecretCapsuleCreateRequestDto) : RetrofitResult<String>
 
-    suspend fun getSecretCapsuleDetail (capsuleId: Int) : RetrofitResult<SecretCapsuleDetail>
+    suspend fun getSecretCapsuleDetail (capsuleId: Long) : RetrofitResult<SecretCapsuleDetail>
 
     suspend fun getSecretCapsuleSummary (capsuleId: Int) : RetrofitResult<SecretCapsuleSummary>
     suspend fun modifySecretCapsule (capsuleId: Int, request: SecretCapsuleModifyRequestDto) : RetrofitResult<SecretCapsuleModify>
